+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 
 ## [Unreleased]
 
+
+## [0.18.0]
+
 - Added: code completion/navigation for //SOURCES
+- Added: Java version synced within module
+- Fixed: external library name always as jbang, and now is `${moduelName}-jbang`
 
 ## [0.17.0]
 - Added: JBang ToolWindow listener to make load JBang script info automatically